Separate LinkEditButton from AttachmentsDropup
